+++ src/main/java/com/github/thedeathlycow/thermoo/api/ThermooRegistries.java
@@ -1,7 +1,6 @@
 package com.github.thedeathlycow.thermoo.api;
 
-import com.github.thedeathlycow.thermoo.api.environment.Environment;
-import com.github.thedeathlycow.thermoo.api.environment.EnvironmentProviderType;
+import com.github.thedeathlycow.thermoo.api.environment.provider.EnvironmentProviderType;
 import com.github.thedeathlycow.thermoo.api.temperature.effects.TemperatureEffect;
 import net.fabricmc.fabric.api.event.registry.FabricRegistryBuilder;
 import net.minecraft.registry.Registry;
