+++ src/main/java/com/github/thedeathlycow/thermoo/api/environment/provider/ConstantEnvironmentProvider.java
@@ -1,7 +1,6 @@
-package com.github.thedeathlycow.thermoo.api.environment;
+package com.github.thedeathlycow.thermoo.api.environment.provider;
 
 import com.github.thedeathlycow.thermoo.api.util.TemperatureRecord;
-import com.mojang.datafixers.util.Either;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
